NOT WORKING TEMP
This commit doesn't work at the moment. I suspect there's something up
with the dependency versions but it's not worth the time it would take
to work it out as I've covered all the course content.

If I come back to this in future - look into the aop dependency. I think
spring-data & spring mvc have clashing dependencies, I've put in an
exclusion but needs further investigation